requirements.txt
azure-functions
azure-storage-common
azure-storage-blob
azure-storage-file-datalake